Crate oak_baremetal_kernel

Main ‘kernel’ for baremetal Oak Functions.

This code takes care of initializing the x86-64 machine properly and handing the reins over to the oak_baremetal_runtime. It is in a separate crate so that we could support different boot protocols (eg Linux boot protocol or PVH) that require different targets, linker scripts and/or provide machine information in different data structures.

Bootloaders (and VMMs under them) have to adhere to the following protocol:

Enter 64-bit long mode, and set up basic paging – enough to load the code, as we will set up a full page table in start_kernel.

Implement a #[panic_handler] that delegates to panic() in this crate.

Call start_kernel from the entry point of the bootloader.

Modules

boot

i8042

Functions

panic

Common panic routine for the kernel. This needs to be wrapped in a panic_handler function in individual bootloader crates.

start_kernel

Main entry point for the kernel, to be called from bootloader.