Trait oak_baremetal_kernel::boot::BootInfo

pub trait BootInfo<E: E820Entry> {
    fn protocol(&self) -> &'static str;
    fn e820_table(&self) -> &[E];
    fn args(&self) -> &CStr;
}

Generic trait for the boot information data structure provided to the kernel from the boot loader.

Required Methods

fn protocol(&self) -> &'static str

Human-readable name of the boot protocol.

fn e820_table(&self) -> &[E]

Slice of address range descriptors representing the memory layout of the machine.

fn args(&self) -> &CStr

Arguments passed to the kernel.